Major refactoring
Used JetBrains to compact a lot of the code.
Changed a lot of the simple functions into lambda functions.
Renamed a ton of variables. Made a lot of variables private, some readonly and const, etc.

If there's a commit that causes issues, it's most likely going to be this one, though I have tried to avoid refactoring important areas for this commit just incase. I'll get to refactoring the rest later.