Reorganising
Just making it easier to see what's going on.
Moved the "WPF" version into the "main" folder, and "archived" the unused work-in-progress version that I was starting with. Was promising, but was far easier and better to work with my existing code, and rewrite from scratch when it works well enough as is.