Porting - Now actually swaps accounts
In this super limited, half moved superposition... it now sucessfully swaps steam accounts! (Right-click - Swap to account)
This is still SUPER basic and breaks if the settings page is uncommented, as things haven't been moved across.

So incredibly happy that I can use a lot of the old XAML UI with the new UI, at least in the meantime.