Steam settings save/load.
The settings page now has persistent settings, even though they don't do anything just yet.
Huge issue though: Previously (being Steam only) there was 1 settings file, now there's going to be 1 per program. This is fine, but it needs to be an object rather than a set-in-stone class. This is fine, just a lot of work to swap out.

A file for program settings (Last size etc).
A file for each switcher in the account switcher.