Refactoring, Set path dialog for game stores
Some refactoring.
Set path dialog for game stores working properly (I think). So you can now set Steam's directory, if it's incorrect.
Main screen is now just a simple image that you click to get to the Steam switcher, as a placeholder. More game stores will go there in the future, as well as a "default" for your favourite program.