Removed old files & commented broken old GUI
Removed SteamFolderInput dialog, as it's already implimented -- But not the check and update every time a character is written in the dialog box.
Removed Globals.cs file, as it just has window controls and things, that are already handled.

Commented out broken GUI (basically everything) from previous version of client, because I'm removing parts that I've added to the new version (Working down to nothing) and removed the window controls etc.