Removing old funcs, fixed modal issue
Removed some old functions I haven't yet removed.
Removed ForgotAccountCheck page, as it's been moved across.

Changed ClearForgottenBackups to just ClearForgotten, as there's the legacy folder, as well as the file with forgotten accounts in it.

Fixed modal buttons being broken on true/false because action was being split differently for an exception, and not split normally after that (WebFuncs.js)

#region Windows Shortcuts seems to be the last thing I need to mvoe out of MainWindow.xaml.cs for it to be fully updated.