Advanced Cleaning + Settings work
Advanced Cleaning disabled until understood checked.
Settings file reverse rule added >> Merges one way or the other, depending on if expecting a file, or not expecting a file. (Merges NEW onto OLD, or OLD onto NEW)
Startup.cs DetailedErrors set to true. I think it'll stay this way in the final build as well, so that it makes debugging somewhat easier for the normal user, and it's open source anyways.