Platform admin update
Further refining of Ubisoft name algorithm.
Platforms can now be run in Admin, and restarted in Admin.
If a platform needs admin to kill: Before doing anything (or even showing platform switcher screen if it's running) and ask user to restart in Admin mode (which should be automatic)
Added new notice dialog popup.