CEFSharp simple. Draggable. Requires another framework...
I DON'T like including yet another runtime (.NET Framework) when it's set to run .NET! This is annoying and unnecessary! I just CAN'T get things to work properly with .NET 5.... The window refuses to drag using info from an example I found. I have more info on JSEvents, so maybe this won't be required in the future. A lot of effort for something that should be simple.

- Added CEFSharp.. Or at least the best I can with .NET 5 for now
- JS-webkit dragging will work with normal WebView2, but not the new "TcNo-CEF" client
- CEFSharp is simple, but draggable. Just need to look into resizing. Somehow forwarding JS to cause motion, like with WebView... But can't find info ANYWHERE.