Steam Web API & Multithread image downloads
- Removed leftover file from Client - Was for loading and displaying screenshots from all games. This is something I want to get to at a later date - A combined screenshot viewer, separated by account then game (if possible).
- Fixed CSS spacing change when input focused or not.
- Added Steam Web API key for faster VAC/Limited status collection and image downloading
- Added multithread image downloader when using Steam Web API key. This should hopefully be expanded to include normal image downloads too (Currently 3 at a time).
- Added time info to verbose console messages.
- Load times for Steam switcher when going back and forth DRASTICALLY improved with Web API - Loads info more efficiently from VAC cache. This needs to be included for non Web API users.

Collection of Steam images seems to work well, as well as VAC info collection. This is definitely an improvement.
IMPORTANT TODO: Also needs some kind of check for Web API key - If invalid: unset.

IMPORTANT NOTE:  This commit works. The Steam switcher works. I was in the middle of updating how each profile is loaded. SaveVacInfo may not work properly for accounts NOT loaded with new Web API. Turns out that on each load, VAC info was being loaded from XML and not the JSON file - Slowing the load a LOT - especially for a lot of accounts. I will finish working on this tomorrow, but for people working on this commit - everything works as normal - Just ignore the ban check and profile system. It's 3 AM, otherwise I would have continued further