Fixed not reloading
There are some serious state issues with this program. I have tried on 3 different commits over ~10 hours to get Instances removed, and Singletons or other DI set up, but efforts are futile with such a large program. It's truly depressing that I can't make this improvement just yet. A lot of cleanup and reorganization needs to happen first, but I'm not sure what to do...

This puts it back in a working state from what I can tell, so at least testing can begin with this, I hope.