Better state management for a few classes.
I am liking this new layout a lot. I can see it improving things, even if they're not named properly. This feels a lot more correct, even if it's more verbose. Should work, I hope. Time to just commit a lot of time as this is going through everything.
What I got wrong about my previous attempts is that there is NO quick way to doing this. Slow and methodical is the way 100%.