Started migration to MAUI
Started migration of Server into just the client. When it's working I need to import everything from the Client -- which isn't much but dragging the window around. Then it should be ready for CLI handling, and release? Sounds good.

This commit does not start currently, but I'm saving as it's late. Lots of work needs to be done here still.