Actually... Never give up.
For ease-of-use and deduplication:
- The program will now install to Local AppData,
- The data will still be located in Roaming AppData.
Updates will no longer need admin, as well as a few other things. Makes life easier for all.

Building the project as self-contained as well should also get rid of the need to download .NET//And make things easier as I don't need to bounce a launch through a check for "Do I have X runtime installed". The WebView error also seems to be greatly improved - Now shows an error when WebView is not installed. Thanks, Microsoft!

Still need to move profile images from UserData to AppData, as well as some other housekeeping before adding window dragging, resizing etc and finally work on the CLI. From that point it's patching everything and it's done... I hope.

- Implemented better handling of missing WebView2. Already automagically should handle missing .NET as all the necessary files are being copied with the program.