2024-07-08 - Skip signing
So, my certificate has expired. I have earned $15-$20 from the lifetime of this project. The certificate I bought a 3-year certificate for $209.00 in 2021. This was OK, but stressful because of having to get attourneys to sign documents to prove identity, etc.
However, now if I were to renew it from one of the cheapest places to get an OV certificate would cost me $657 + $130 for a hardware token, then there's shipping to South Africa too. Absolutely insane.

I am looking for alternatives, but a build or two may come out with no certificate, unfortunately.