2024-08-24 - Added missing Steam loginusers option
No idea when this was added, nor what it really does. Defaults to 1, like Steam does (if missing Steam adds this to every account on startup) and will keep the value if the user changes it.
Settings in this file need to be explicitly named, otherwise they are forgotten when changing accounts.