2024-09-02 - Moved Platforms.json to UserData
User data folder can be modified when the TcNo Account Switcher is installed, which it can't be when the normal program is installed and NOT running as admin. The user data folder is a much better location for this file. Will still appear in userdata folder when installing, so it is moved after installing.